add long data type; currently used only by Take and Drop. Probably should also be supported by TakeRight, Length, ...

+++ src/main/scala/de/sciss/fscape/graph/Drop.scala
@@ -27,6 +27,6 @@ final case class Drop(in: GE, len: GE) extends UGenSource.SingleOut {
 
   private[fscape] def makeStream(args: Vec[StreamIn])(implicit b: stream.Builder): StreamOut = {
     val Vec(in, len) = args
-    stream.Drop(in = in.toDouble, len = len.toInt)
+    stream.Drop(in = in.toDouble, len = len.toLong)
   }
 }
